# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API калькулятора с кэшированием, хранением истории и аналитикой.\nПоддерживает операции: сложение (+), вычитание (-), умножение (*), деление (/).",
        "title": "LizzyCalc API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/calculate": {
            "post": {
                "description": "Принимает два числа и операцию (+, -, *, /), возвращает результат. Результат кэшируется и сохраняется в БД.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Выполнить вычисление",
                "parameters": [
                    {
                        "description": "Параметры вычисления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат вычисления",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.CalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос или неизвестная операция",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.CalculateResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.CalculateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/history": {
            "get": {
                "description": "Возвращает список всех выполненных операций из БД",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Получить историю операций",
                "responses": {
                    "200": {
                        "description": "Список операций",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.HistoryResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_controllers_calculator.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/liveness": {
            "get": {
                "description": "Возвращает статус \"alive\" если сервер запущен",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Проверка жизнеспособности",
                "responses": {
                    "200": {
                        "description": "status: alive",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/readyness": {
            "get": {
                "description": "Проверяет подключение к БД и возвращает статус готовности",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Проверка готовности",
                "responses": {
                    "200": {
                        "description": "status: ready",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "status: not ready, error: ...",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api_http_controllers_calculator.CalculateRequest": {
            "description": "Параметры для выполнения арифметической операции",
            "type": "object",
            "required": [
                "number1",
                "number2",
                "operation"
            ],
            "properties": {
                "number1": {
                    "type": "number",
                    "example": 10.5
                },
                "number2": {
                    "type": "number",
                    "example": 5.2
                },
                "operation": {
                    "type": "string",
                    "enum": [
                        "+",
                        "-",
                        "*",
                        "/"
                    ],
                    "example": "+"
                }
            }
        },
        "internal_api_http_controllers_calculator.CalculateResponse": {
            "description": "Результат арифметической операции",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "division by zero"
                },
                "result": {
                    "type": "number",
                    "example": 15.7
                }
            }
        },
        "internal_api_http_controllers_calculator.ErrorResponse": {
            "description": "Сообщение об ошибке",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
        "internal_api_http_controllers_calculator.HistoryItem": {
            "description": "Информация об одной выполненной операции",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "type": "string",
                    "example": ""
                },
                "number1": {
                    "type": "number",
                    "example": 10.5
                },
                "number2": {
                    "type": "number",
                    "example": 5.2
                },
                "operation": {
                    "type": "string",
                    "example": "+"
                },
                "result": {
                    "type": "number",
                    "example": 15.7
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                }
            }
        },
        "internal_api_http_controllers_calculator.HistoryResponse": {
            "description": "Список всех выполненных операций",
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_http_controllers_calculator.HistoryItem"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "description": "Операции калькулятора: вычисления и история",
            "name": "calculator"
        },
        {
            "description": "Системные эндпоинты: liveness, readiness",
            "name": "system"
        }
    ]
}